@@ -12,16 +12,16 @@
   padding={[0, 0, 0, 0]}
   wrap={true}
 >
-  {#each values as value, index (index)}
+  {#each values as _, index (index)}
     <TextField
       bind:this={textFields[index]}
       max={viewModel.max}
       min={viewModel.min}
+      onBlur={textFieldBlurHandler}
+      onFocus={() => { textFieldFocusHandler(index) }}
       type="number"
       width={40}
-      bind:value={value}
-      on:blur={textFieldBlurHandler}
-      on:focus={() => { textFieldFocusHandler(index) }}
+      bind:value={values[index]}
     />
   {/each}
 </Stack>
@@ -33,21 +33,20 @@
   import { colors } from "@scripts/colors"
   import type { IntegerInputGroupViewModel } from "@shared/types/viewModels"
   import { isNotNullish, isNullish } from "@shared/utils"
-  import { onMount } from "svelte"
   
-  export let viewModel: IntegerInputGroupViewModel
+  type Props = {
+    viewModel: IntegerInputGroupViewModel
+  }
   
-  const textFields: TextField<number, "number">[] = []
+  const {
+    viewModel = $bindable(),
+  }: Props = $props()
+  
+  const textFields: TextField<number, "number">[] = $state([])
   
   let oldValues: number[] = []
-  let values: number[] = []
+  let values: number[] = $state([])
   let currentIndex: number | null = null
-  let hasMounted = false
-  let error: string | null = null
-  
-  onMount(() => {
-    hasMounted = true
-  })
 
   const textFieldFocusHandler = (index: number) => {
     if (isNullish(currentIndex)) {
@@ -60,7 +59,28 @@
   const textFieldBlurHandler = () => {
     requestAnimationFrame(() => {
       if (isNullish(textFields.find((textField) => { return textField.isFocused() }))) {
-        if (isNotNullish(error)) {
+        let error: string | null = null
+    
+        values.forEach((value) => {
+          if (Number.isNaN(value)) {
+            error = `All ${viewModel.name} values are required.`
+          } else if (!Number.isInteger(value)) {
+            error = `All ${viewModel.name} values must be integers.`
+          } else if (isNotNullish(viewModel.min) && value < viewModel.min) {
+            error = `All ${viewModel.name} values must be greater than or equal to ${viewModel.min}.`
+          } else if (isNotNullish(viewModel.max) && value > viewModel.max) {
+            error = `All ${viewModel.name} values must be less than or equal to ${viewModel.max}.`
+          }
+        })
+    
+        if (isNullish(error) && isNotNullish(viewModel.sum) && values.reduce((result, value) => { return result + value }, 0) !== viewModel.sum) {
+          error = `All ${viewModel.name} values must add up to ${viewModel.sum}.`
+        }
+    
+        if (isNullish(error)) {
+          viewModel.values = values.map((value) => { return value })
+          error = null
+        } else {
           values = oldValues
           showErrorDialog(error)
         }
@@ -69,43 +89,8 @@
       }
     })
   }
-
-  $: values, valuesListener()
-  const valuesListener = () => {
-    if (!hasMounted) {
-      return
-    }
-    
-    let foundNewError = false
-    
-    values.forEach((value) => {
-      if (Number.isNaN(value)) {
-        error = `All ${viewModel.name} values are required.`
-        foundNewError = true
-      } else if (!Number.isInteger(value)) {
-        error = `All ${viewModel.name} values must be integers.`
-        foundNewError = true
-      } else if (isNotNullish(viewModel.min) && value < viewModel.min) {
-        error = `All ${viewModel.name} values must be greater than or equal to ${viewModel.min}.`
-        foundNewError = true
-      } else if (isNotNullish(viewModel.max) && value > viewModel.max) {
-        error = `All ${viewModel.name} values must be less than or equal to ${viewModel.max}.`
-        foundNewError = true
-      }
-    })
-    
-    if (!foundNewError && isNotNullish(viewModel.sum) && values.reduce((result, value) => { return result + value }, 0) !== viewModel.sum) {
-      error = `All ${viewModel.name} values must add up to ${viewModel.sum}.`
-      foundNewError = true
-    }
-    
-    if (!foundNewError) {
-      viewModel.values = values.map((value) => { return value })
-      error = null
-    }
-  }
-
-  $: viewModel.values, viewModelValuesListener()
+  
+  $effect(() => { viewModel.values; viewModelValuesListener() })
   const viewModelValuesListener = () => {
     values = viewModel.values.map((value) => { return value })
   }
