@@ -1,30 +1,29 @@
 <TextField
   max={viewModel.max}
   min={viewModel.min}
+  onBlur={textFieldBlurHandler}
+  onFocus={textFieldFocusHandler}
   title={viewModel.name}
   type="number"
   bind:value={value}
-  on:blur={textFieldBlurHandler}
-  on:focus={textFieldFocusHandler}
 />  <!-- TODO: Description -->
 <script lang="ts">
   import { showErrorDialog } from "@components/dialogs/DialogContainer.svelte"
   import TextField from "@components/inputs/TextField.svelte"
   import type { IntegerInputViewModel } from "@shared/types/viewModels"
   import { isNotNullish, isNullish } from "@shared/utils"
-  import { onMount } from "svelte"
   
-  export let viewModel: IntegerInputViewModel
+  type Props = {
+    viewModel: IntegerInputViewModel
+  }
+  
+  const {
+    viewModel = $bindable(),
+  }: Props = $props()
   
   let oldValue: number | undefined = undefined
-  let value: number | undefined = undefined
-  let hasMounted = false
+  let value: number | undefined = $state(viewModel.value)
   let error: string | null = null
-  
-  onMount(() => {
-    value = viewModel.value ?? Number.NaN
-    hasMounted = true
-  })
   
   const textFieldFocusHandler = () => {
     oldValue = value
@@ -32,18 +31,6 @@
   }
   
   const textFieldBlurHandler = () => {
-    if (isNotNullish(error)) {
-      value = oldValue
-      showErrorDialog(error)
-    }
-  }
-  
-  $: value, valueListener()
-  const valueListener = () => {
-    if (!hasMounted) {
-      return
-    }
-    
     if (Number.isNaN(value) || isNullish(value)) {
       if (viewModel.isRequired) {
         error = `${viewModel.name} is required.`
@@ -60,10 +47,15 @@
       error = null
       viewModel.value = value
     }
+      
+    if (isNotNullish(error)) {
+      value = oldValue
+      showErrorDialog(error)
+    }
   }
   
-  $: viewModel.value, configValueListener()
-  const configValueListener = () => {
+  $effect(() => { viewModel.value; viewModelValueListener() })
+  const viewModelValueListener = () => {
     value = viewModel.value
   }
 </script>
