@@ -1,4 +1,9 @@
 <script lang="ts">
-  export let neverValue: never
-  neverValue
+  type Props = {
+    neverValue: never
+  }
+  
+  const {
+    neverValue, // eslint-disable-line @typescript-eslint/no-unused-vars
+  }: Props = $props()
 </script>
